Fix a control-GNN crash under AMP, and add a queue runner for sequential arms
The control GNN had never been executed. `wnca train -c configs/phase2d_control.yaml`
crashes on its first training step:

  RuntimeError: index_add_(): self (Float) and source (Half) must have the same scalar type

`index_add_` raises on mismatched scalar types rather than promoting. On CUDA, autocast
sends the edge MLP's output to half/bf16 while LayerNorm hands `h` back in fp32, so from
the second message-passing layer onward the fp32 accumulator meets a half source. The
accumulator has to take the MESSAGE dtype. Caught by smoke-testing the config before
committing a queue to it; it would otherwise have killed an overnight cloud run on the
first step.

The regression test has to be a CUDA test: CPU autocast keeps LayerNorm in bf16, so the
dtypes agree there and a CPU version passes with or without the fix. Verified by reverting
the fix -- fails in both fp16 and bf16, passes with it.

Also adds test_phase2d_control_is_parameter_matched_to_phase2c, which asserts the two real
configs stay within 10% of each other on parameter count. The shipped config was 1.67x and
nothing caught it; now an edit that breaks the match fails a test instead of silently
producing an uninterpretable experiment.

scripts/run_queue.sh runs jobs back-to-back on one instance and stops it only after the
last one. autostop.sh is right for a single run and wrong for a queue: with GPU quota at 1
the 2d arms must share an instance, and a shutdown between them would need a manual restart
into a possible capacity stockout each time. It waits for any in-flight run first, so it
can be armed against a run that is already going, and a failing job does not skip the
remaining jobs or the shutdown.

175 tests pass.

+++ src/wnca/models/control_gnn.py
@@ -61,7 +61,12 @@ class MessagePassingLayer(nn.Module):
 
     def forward(self, h: torch.Tensor) -> torch.Tensor:  # [B, N, H]
         m = self.edge_mlp(torch.cat([h[:, self.src], h[:, self.dst]], dim=-1))  # [B, E, H]
-        agg = torch.zeros_like(h).index_add_(1, self.dst, m) * self.inv_deg
+        # The accumulator must match the MESSAGE dtype, not `h`'s. Under autocast the edge MLP
+        # returns half/bf16 while `h` comes back from LayerNorm in fp32, and `index_add_`
+        # requires both operands to share a scalar type -- it raises rather than promoting.
+        # Multiplying by the fp32 `inv_deg` then promotes the mean back to fp32, which is where
+        # the aggregation wants to be anyway.
+        agg = torch.zeros_like(h, dtype=m.dtype).index_add_(1, self.dst, m) * self.inv_deg
         return self.norm(h + self.node_mlp(torch.cat([h, agg], dim=-1)))
